refactored furnish.js and all seeds so that venues will persist but events will be renewed every night
