+++ package.json
@@ -12,7 +12,9 @@
     "db:drop": "node src/api/database/drop.js",
     "db:drop:test": "NODE_ENV=test npm run db:drop",
     "db:furnish": "node src/api/database/furnish",
+    "db:nightly-script": "node src/api/database/nightlyScript",
     "db:seed:admins": "node src/api/database/seedAdmins",
+    "db:seed:venues": "node src/api/database/seedVenues",
     "lint": "eslint ./src",
     "server:start": "nodemon --watch src/api src/api/server.js",
     "server:test": "npm run db:drop:test && NODE_ENV=test mocha --opts 'support/mocha.opts' 'src/api/**/*.spec.js'",
