+++ src/api/database/index.js
@@ -4,16 +4,15 @@ const mongoose = require('mongoose');
 mongoose.Promise = global.Promise;
 
 const config = require('../config');
-// This will make a comeback in the future.
-// const isEnvironment = require('../utils/isEnvironment');
+const isEnvironment = require('../utils/isEnvironment');
 const log = require('../utils/log');
 
 const database = {
   drop() {
-    return new Promise((resolve) => {
-      // if (isEnvironment(['production', 'preproduction'])) {
-      //   reject(log('ERROR: production and preproduction databases cannot be dropped'));
-      // }
+    return new Promise((resolve, reject) => {
+      if (isEnvironment(['production', 'preproduction'])) {
+        reject(log('ERROR: production and preproduction databases cannot be dropped'));
+      }
 
       mongoose.connection.dropDatabase(() => {
         resolve(log('DATABASE CLEANED'));
