+++ src/api/database/seeds/venues/apex/abbeyRoad.js
@@ -1,0 +1,13 @@
+'use strict';
+
+const venue = {
+  name: 'Abbey Road Tavern & Grill - Apex',
+  address: '1700 Center St',
+  city: 'Apex',
+  state: 'NC',
+  zipCode: 27502,
+  latitude: 35.725996,
+  longitude: -78.827330
+};
+
+module.exports = venue;
