+++ src/api/database/seeds/venues/cary/abbeyRoad.js
@@ -1,0 +1,13 @@
+'use strict';
+
+const venue = {
+  name: 'Abbey Road Tavern & Grill - Cary',
+  address: '1195 W Chatham St',
+  city: 'Cary',
+  state: 'NC',
+  zipCode: 27513,
+  latitude: 35.770850,
+  longitude: -78.807000
+};
+
+module.exports = venue;
