+++ src/api/database/seeds/venues/durham/beyuCaffe.js
@@ -1,0 +1,13 @@
+'use strict';
+
+const venue = {
+  address: '341 W Main St',
+  city: 'Durham',
+  latitude: 35.996665,
+  longitude: -78.903873,
+  name: 'Beyú Caffe',
+  state: 'NC',
+  zipCode: 27701
+};
+
+module.exports = venue;
