+++ src/api/database/seeds/venues/durham/blueNoteGrill.js
@@ -1,0 +1,13 @@
+'use strict';
+
+const venue = {
+  name: 'Blue Note Grill',
+  address: '709 Washington Street',
+  city: 'Durham',
+  state: 'NC',
+  zipCode: 27704,
+  latitude: 36.004379,
+  longitude: -78.902946
+};
+
+module.exports = venue;
