+++ src/api/database/seeds/venues/durham/bottle501.js
@@ -1,0 +1,13 @@
+'use strict';
+
+const venue = {
+  address: '3219 Watkins Rd',
+  city: 'Durham',
+  latitude: 35.953073,
+  longitude: -78.992012,
+  name: 'Bottle 501',
+  state: 'NC',
+  zipCode: 27707
+};
+
+module.exports = venue;
