+++ src/api/database/seeds/venues/durham/skewersBarAndGrill.js
@@ -1,0 +1,13 @@
+'use strict';
+
+const venue = {
+  address: '1013 W Main St',
+  city: 'Durham',
+  latitude: 36.001054,
+  longitude: -78.911168,
+  name: 'Skewers Bar and Grill',
+  state: 'NC',
+  zipCode: 27701
+};
+
+module.exports = venue;
