+++ src/api/database/seeds/venues/durham/socialDurham.js
@@ -1,0 +1,13 @@
+'use strict';
+
+const venue = {
+  name: 'Social Durham',
+  address: '1007 W Main St',
+  city: 'Durham',
+  state: 'NC',
+  zipCode: 27701,
+  latitude: 36.000652,
+  longitude: -78.910782
+};
+
+module.exports = venue;
