+++ src/api/database/seeds/venues/durham/thePinhook.js
@@ -1,0 +1,13 @@
+'use strict';
+
+const venue = {
+  address: '115 W Main St',
+  city: 'Durham',
+  latitude: 35.995280,
+  longitude: -78.901512,
+  name: 'The Pinhook',
+  state: 'NC',
+  zipCode: 27701
+};
+
+module.exports = venue;
