+++ src/api/database/seeds/venues/index.js
@@ -1,0 +1,12 @@
+module.exports = [
+  require('./apex/abbeyRoad'),
+  require('./cary/abbeyRoad'),
+  require('./durham/beyuCaffe'),
+  require('./durham/blueNoteGrill'),
+  require('./durham/bottle501'),
+  require('./durham/skewersBarAndGrill'),
+  require('./durham/socialDurham'),
+  require('./raleigh/deepSouth'),
+  require('./raleigh/fallsRiverMusic'),
+  require('./raleigh/gizmoBrewWorks')
+];
