+++ src/api/database/seeds/venues/raleigh/deepSouth.js
@@ -1,0 +1,13 @@
+'use strict';
+
+const venue = {
+  name: 'Deep South The Bar',
+  address: '430 S Dawson St',
+  city: 'Raleigh',
+  state: 'NC',
+  zipCode: 27601,
+  latitude: 35.774731,
+  longitude: -78.643929
+};
+
+module.exports = venue;
