+++ src/api/database/seeds/venues/raleigh/fallsRiverMusic.js
@@ -1,0 +1,14 @@
+'use strict';
+
+const venue = {
+  name: 'Falls River Music',
+  address: '10930 Raven Ridge Rd',
+  city: 'Raleigh',
+  state: 'NC',
+  zipCode: 27614,
+  latitude: 35.907351,
+  longitude: -78.590152
+};
+
+module.exports = venue;
+
