+++ src/api/database/seeds/venues/raleigh/gizmoBrewWorks.js
@@ -1,0 +1,14 @@
+'use strict';
+
+const venue = {
+  name: 'Gizmo Brew Works',
+  address: '5907 Triangle Dr',
+  city: 'Raleigh',
+  state: 'NC',
+  zipCode: 27617,
+  latitude: 35.896730,
+  longitude: -78.744105
+};
+
+module.exports = venue;
+
