+++ src/api/lib/populator/index.js
@@ -1,13 +1,16 @@
 'use strict';
 
+const _ = require('lodash');
+
+const buildEventPromises = require('../../utils/buildEventPromises');
 const log = require('../../utils/log');
-const saveEventsAndVenue = require('../../utils/saveEventsAndVenue');
+const Event = require('../../app/models/event');
 
 const populator = {
   addMicNights(sites, message) {
     return Promise
       .all(sites)
-      .then((eventsAndVenues) => populator.saveMicNights(eventsAndVenues))
+      .then((events) => populator.renewMicNights(events))
       .then(() => log(message))
       .catch((err) => {
         // no longer throwing here because on fail, I don't need to the whole operation to stop
@@ -15,14 +18,20 @@ const populator = {
       });
   },
 
-  saveMicNights(eventsAndVenues) {
-    const filteredArray = eventsAndVenues.filter(foo => foo);
+  renewMicNights(events) {
+    const handsomeEvents = _.compact(_.flatten(events));
+    const venueIds = handsomeEvents.map((event) => event.venue);
 
-    const promises = filteredArray.map(({ events, venue }) => {
-      return saveEventsAndVenue(events, venue);
-    });
+    return Event
+      .remove({ venue: { $in: venueIds } })
+      .then(() => {
+        const promises = handsomeEvents.reduce((memo, event) => {
+          if (event) { memo.push(buildEventPromises(event)); }
+          return memo;
+        }, []);
 
-    return Promise.all(promises);
+        return Promise.all(_.flatten(promises));
+      });
   }
 };
