+++ src/api/lib/siteScraper/index.js
@@ -15,22 +15,23 @@ const sites = [
   require('./sites/raleigh/gizmoBrewWorks')
 ];
 
-function findEventsAndVenue(site, resolve, $) {
-  try {
-    const { events, venue } = site.run($);
+function buildEventsFromSite(site, resolve, $) {
+  return site
+    .run($)
+    .then((events) => {
+      if (events.length < 1) {
+        emailJordanAboutHisBadScraper(site, { stack: 'No events were found.' });
+        // resolving because I just want this to return null so other sites
+        // that are being scraped can get their events displayed
+        return resolve();
+      }
 
-    if (events.length < 1) {
-      emailJordanAboutHisBadScraper(site, { stack: 'No events were found.' });
-      // resolving because I just want this to return null so other sites
-      // that are being scraped can get their events displayed
-      resolve();
-    } else {
-      resolve({ events, venue });
-    }
-  } catch (e) {
-    emailJordanAboutHisBadScraper(site, e);
-    resolve();
-  }
+      return resolve(events);
+    })
+    .catch((e) => {
+      emailJordanAboutHisBadScraper(site, e);
+      return resolve();
+    });
 }
 
 function scrape(site, error, html, resolve, reject) {
@@ -41,15 +42,15 @@ function scrape(site, error, html, resolve, reject) {
   // htmlparser2 is what's used in the cheerio library
   if (!htmlparser.parseDOM(html)[0]) {
     const message = `${site.url} did not return any html`;
-    console.log('html of bad site:', html);
-    console.error(message);
+    console.error(message); // eslint-disable-line no-console
+    console.log('html of bad site:', html); // eslint-disable-line no-console
     emailJordanAboutHisBadScraper(site, { stack: message });
     return resolve();
   }
 
   const $ = cheerio.load(html);
 
-  return findEventsAndVenue(site, resolve, $);
+  return buildEventsFromSite(site, resolve, $);
 }
 
 const siteScraper = {
