+++ src/api/lib/siteScraper/sites/apex/abbeyRoad.js
@@ -3,35 +3,28 @@
 const moment = require('moment');
 const searchForOpenMicNights = require('../../../../utils/searchForOpenMicNights');
 const url = 'http://abbeyroadnc.com/open-mic/';
+const Venue = require('../../../../app/models/venue');
 
 function abbeyRoad($) {
-  const micNights = searchForOpenMicNights('OPEN MIC - APEX', $, 'p');
-  const noMoreTuesdays = $(micNights[0]).text().indexOf('Tuesday') < 0;
+  const events = (venue) => {
+    const micNights = searchForOpenMicNights('OPEN MIC - APEX', $, 'p');
+    const noMoreTuesdays = $(micNights[0]).text().indexOf('Tuesday') < 0;
 
-  if (noMoreTuesdays) { throw new Error('Abbey Road - Apex no longer has open mic nights on Tuesdays'); }
+    if (noMoreTuesdays) { throw new Error('Abbey Road - Apex no longer has open mic nights on Tuesdays'); }
 
-  const events = [{
-    frequency: 'weekly',
-    startDate: moment().isoWeekday(2).format('MM-DD-YYYY'),
-    startTime: '8pm - 11pm',
-    title: 'OPEN MIC - APEX',
-    type: 'open'
-  }];
-
-  const venue = {
-    name: 'Abbey Road Tavern & Grill - Apex',
-    address: '1700 Center St',
-    city: 'Apex',
-    state: 'NC',
-    zipCode: 27502,
-    latitude: 35.725996,
-    longitude: -78.827330
+    return [{
+      frequency: 'weekly',
+      startDate: moment().isoWeekday(2).format('MM-DD-YYYY'),
+      startTime: '8pm - 11pm',
+      title: 'OPEN MIC - APEX',
+      type: 'open',
+      venue: venue._id
+    }];
   };
 
-  return {
-    events,
-    venue
-  };
+  return Venue
+    .findOne({ name: 'Abbey Road Tavern & Grill - Apex' })
+    .then((venue) => events(venue));
 }
 
 module.exports = {
