+++ src/api/lib/siteScraper/sites/cary/abbeyRoad.js
@@ -3,35 +3,28 @@
 const moment = require('moment');
 const searchForOpenMicNights = require('../../../../utils/searchForOpenMicNights');
 const url = 'http://abbeyroadnc.com/open-mic/';
+const Venue = require('../../../../app/models/venue');
 
 function abbeyRoad($) {
-  const micNights = searchForOpenMicNights('OPEN MIC - CARY', $, 'p');
-  const noMoreWednesdays = $(micNights[0]).text().indexOf('Wednesday') < 0;
+  const events = (venue) => {
+    const micNights = searchForOpenMicNights('OPEN MIC - CARY', $, 'p');
+    const noMoreWednesdays = $(micNights[0]).text().indexOf('Wednesday') < 0;
 
-  if (noMoreWednesdays) { throw new Error('Abbey Road - Cary no longer has open mic nights on Wednesdays'); }
+    if (noMoreWednesdays) { throw new Error('Abbey Road - Cary no longer has open mic nights on Wednesdays'); }
 
-  const events = [{
-    frequency: 'weekly',
-    startDate: moment().isoWeekday(3).format('MM-DD-YYYY'),
-    startTime: '9pm - 12am',
-    title: 'OPEN MIC - CARY',
-    type: 'open'
-  }];
-
-  const venue = {
-    name: 'Abbey Road Tavern & Grill - Cary',
-    address: '1195 W Chatham St',
-    city: 'Cary',
-    state: 'NC',
-    zipCode: 27513,
-    latitude: 35.770850,
-    longitude: -78.807000
+    return [{
+      frequency: 'weekly',
+      startDate: moment().isoWeekday(3).format('MM-DD-YYYY'),
+      startTime: '9pm - 12am',
+      title: 'OPEN MIC - CARY',
+      type: 'open',
+      venue: venue._id
+    }];
   };
 
-  return {
-    events,
-    venue
-  };
+  return Venue
+    .findOne({ name: 'Abbey Road Tavern & Grill - Cary' })
+    .then((venue) => events(venue));
 }
 
 module.exports = {
