+++ src/api/lib/siteScraper/sites/durham/socialDurham.js
@@ -2,6 +2,7 @@
 
 const createWeeklyDates = require('../../../../utils/createWeeklyDates');
 const url = 'http://socialdurham.com/events/';
+const Venue = require('../../../../app/models/venue');
 
 function getMicNight($) {
   return $('#text-3 .textwidget')
@@ -25,29 +26,21 @@ function socialDurham($) {
   const startDates = createWeeklyDates(micNightDay);
   const startTime = micNight[1].split(' ')[0];
 
-  const events = startDates.map((startDate) => {
-    return {
-      startDate,
-      startTime,
-      title: 'Song writers open mic',
-      type: 'open'
-    };
-  });
-
-  const venue = {
-    name: 'Social Durham',
-    address: '1007 W Main St',
-    city: 'Durham',
-    state: 'NC',
-    zipCode: 27701,
-    latitude: 36.000652,
-    longitude: -78.910782
+  const events = (venue) => {
+    return startDates.map((startDate) => {
+      return {
+        startDate,
+        startTime,
+        title: 'Song writers open mic',
+        type: 'open',
+        venue: venue._id
+      };
+    });
   };
 
-  return {
-    events,
-    venue
-  };
+  return Venue
+    .findOne({ name: 'Social Durham' })
+    .then((venue) => events(venue));
 }
 
 module.exports = {
