+++ src/api/lib/siteScraper/sites/raleigh/fallsRiverMusic.js
@@ -3,32 +3,23 @@
 const moment = require('moment');
 const searchForOpenMicNights = require('../../../../utils/searchForOpenMicNights');
 const url = 'http://www.fallsrivermusic.com/';
+const Venue = require('../../../../app/models/venue');
 
 function fallsRiverMusic($) {
-  searchForOpenMicNights('Every WEDNESDAY Night 8pm – 10pm', $, 'p');
+  return Venue
+    .findOne({ name: 'Falls River Music' })
+    .then((venue) => {
+      searchForOpenMicNights('Every WEDNESDAY Night 8pm – 10pm', $, 'p');
 
-  const events = [{
-    frequency: 'weekly',
-    startDate: moment().isoWeekday(3).format('MM-DD-YYYY'),
-    startTime: '8 - 10pm',
-    title: 'OPEN-MIC NIGHT',
-    type: 'open'
-  }];
-
-  const venue = {
-    name: 'Falls River Music',
-    address: '10930 Raven Ridge Rd',
-    city: 'Raleigh',
-    state: 'NC',
-    zipCode: 27614,
-    latitude: 35.907351,
-    longitude: -78.590152
-  };
-
-  return {
-    events,
-    venue
-  };
+      return [{
+        frequency: 'weekly',
+        startDate: moment().isoWeekday(3).format('MM-DD-YYYY'),
+        startTime: '8 - 10pm',
+        title: 'OPEN-MIC NIGHT',
+        type: 'open',
+        venue: venue._id
+      }];
+    });
 }
 
 module.exports = {
